when setting config keys, validate against struct before writing to disk

+++ repo/config/log.go
@@ -2,7 +2,7 @@ package config
 
 
 type Log struct {
-	MaxSizeMB  uint64
-	MaxBackups uint64
-	MaxAgeDays uint64
+	MaxSizeMB  int
+	MaxBackups int
+	MaxAgeDays int
 }
